requirements.txt
# Core AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# OSINT Dependencies
shodan>=1.28.0
beautifulsoup4>=4.11.0
requests>=2.28.0
python-whois>=0.7.3
dnspython>=2.2.0
phonenumbers>=8.12.0
pillow>=9.2.0
pypdf2>=2.10.0
fpdf>=1.7.2
python-dotenv>=0.19.0

# LLM Integration
openai>=1.0.0
anthropic>=0.25.0
langchain>=0.1.0
langchain-community>=0.0.20

# Neural Network Training
tensorflow>=2.13.0
keras>=2.13.0
pytorch-lightning>=2.0.0

# Security & Web Tools
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
python-whois>=0.8.0
shodan>=1.30.0
python-nmap>=0.7.0

# URL/Domain Analysis
urlparse3>=1.1
tldextract>=3.4.0
validators>=0.20.0

# Data Processing
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Visualization
plotly>=5.15.0
wordcloud>=1.9.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0
configparser>=5.3.0

# CLI & Interface
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Report Generation
fpdf2>=2.7.0
markdown>=3.4.0
jinja2>=3.1.0

# Async Support
aiohttp>=3.8.0
asyncio>=3.4.3
